Update Docker-compose configuration for evento-server
The Docker-compose file used for running the evento-server and evento-db has been updated. This includes changes to always restart the postgres service, use the latest image of the evento-server, and add several environment variables for the evento-server service. Some unused lines at the end of the file have also been removed for cleanliness.